:bug: fix emojis route
